Bugs fixed
- Now, for every db save reuqest ,page will not be refreshed.
- Added evaluation route implementation.
- Input checking now will be made before "save" btn press.

+++ src/app/admin-page/admin-page.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { DBService } from '../services/db.service';
+import { RegisterLoginService } from '../services/register-login.service';
 
 
 @Component({
@@ -14,52 +15,72 @@ export class AdminPageComponent implements OnInit {
   showChangeQuestionForm = [];
   addNewQuestionFormToggle = false;
 
-  constructor(private db: DBService) {
+  constructor(private db: DBService, private auth: RegisterLoginService) {
+    auth.IsAuth = localStorage.loggedUser ? true : false;
     if (db.isAdmin) {
-      this.db.getAllUsersFromDb().subscribe(
-        (users) => {
-          this.DBUsers.push(users);
-          this.DBUsers = this.DBUsers[0];
-          this.DBUsers.shift();
-        }
-      )
-      this.db.getAllQuestionsFromDB().subscribe(
-        (questions) => {
-          questions.shift();
-          this.Questions.push(...questions);
-          for (let i = 0; i < questions.lenght; i++) {
-            this.showChangeQuestionForm[i] = false;
+      if (this.DBUsers.length == 0) {
+        this.db.getAllUsersFromDb().subscribe(
+          (users) => {
+            this.DBUsers.push(users);
+            this.DBUsers = this.DBUsers[0];
+            this.DBUsers.shift();
           }
-        }
-      )
+        )
+      }
+      if (this.Questions.length == 0) {
+        this.db.getAllQuestionsFromDB().subscribe(
+          (questions) => {
+            questions.shift();
+            this.Questions.push(...questions);
+            for (let i = 0; i < questions.lenght; i++) {
+              this.showChangeQuestionForm[i] = false;
+            }
+          }
+        )
+      }
     }
   }
-  changeVolunteerStatusWithId(statusData, userId) {
+
+  changeVolunteerStatusWithId(statusData, userId, vectorIndex) {
     this.db.changeUserStatus(statusData, userId);
+    this.DBUsers[vectorIndex].userStatus = statusData.newStatus;
   }
-  saveEditedQuestion(newQuestion, id) {
-    this.db.editQuestion(newQuestion, id);
+
+  saveEditedQuestion(newQuestion, id, index) {
+    if (this.db.checkTextlength(newQuestion.content)) {
+      this.db.editQuestion(newQuestion, id);
+      this.Questions[index].content = newQuestion.content;
+      this.toggleEditInput(index);
+    }
   }
 
   toggleEditInput(index) {
     this.showChangeQuestionForm[index] = !this.showChangeQuestionForm[index];
   }
+
   toggleAddNewQuestionBtn() {
     this.addNewQuestionFormToggle = !this.addNewQuestionFormToggle;
   }
-  deleteUser(userId) {
-    this.db.deleteUser(userId);
-  }
-  editQuestion(id) {
-    console.log(id);
-  }
-  saveNewQuestion(newQuestion) {
-    this.db.saveNewQuestion(newQuestion);
-  }
-  deleteQuestion(id) {
-    this.db.deleteQuestion(id);
+
+  deleteUser(userId, index, userName) {
+    if (confirm("Are you sure you want to delete " + userName + " ?")) {
+      this.db.deleteUser(userId);
+      this.DBUsers.splice(index, 1);
+    }
   }
 
+  saveNewQuestion(newQuestion) {
+    if (this.db.checkTextlength(newQuestion.content)) {
+      this.db.saveNewQuestion(newQuestion);
+      this.Questions.push(newQuestion);
+      this.addNewQuestionFormToggle = false;
+    }
+  }
+
+  deleteQuestion(id, index) {
+    this.db.deleteQuestion(id);
+    this.Questions.splice(index, 1);
+  }
 
   ngOnInit() {
   }
